termios: check to see if termios speed_t is a direct map to baud
If termios speed_t is a direct mapping to bauds, it is presumably safe
to assume that it can be used as a generic interface.  This applies to
Linux with glibc 2.42+, GNU Hurd, and at least some BSDs.

Try to detect this case and if so, do the simple thing.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>